+++ main.go
@@ -1,8 +1,11 @@
 package main
 
+import sitemap_builder "goPhercise/sitemap-builder"
+
 func main() {
 	//quiz_game.RollQuizGame()
 	//url_shortener.RollURLShortener()
 	//cya_game.RollCYAGame()
 	//HTML Link parser is run in its own directories
+	sitemap_builder.RollSitemapBuilder()
 }